+++ CHANGES.rst
@@ -29,6 +29,8 @@ Mosviz
 Specviz
 ^^^^^^^
 
+- Exposed toggle in Plot Options plugin for viewing uncertainties. [#1189]
+
 API Changes
 -----------